Fix narrowing error in MCCAS in 32 bit platforms
(cherry picked from commit 9a3f96a40bbfb473f517eda9afa2be39596ab27f)

+++ llvm/lib/MCCAS/MCCASObjectV1.cpp
@@ -104,7 +104,7 @@ public:
 };
 
 struct CUInfo {
-  size_t CUSize;
+  uint64_t CUSize;
   uint32_t AbbrevOffset;
 };
 static Expected<CUInfo> getAndSetDebugAbbrevOffsetAndSkip(
